Prevent layer context menus to be polluted by JupyterLab
