+++ packages/base/src/workspace/panels/components/layers.tsx
@@ -2,15 +2,19 @@ import { faEllipsisVertical } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   IJGISLayerGroup,
+  IJGISLayerItem,
   IJGISLayerTree,
   IJupyterGISModel,
+  IJupyterGISWidget,
   ISelection,
+  ProcessingMerge,
   SelectionType,
 } from '@jupytergis/schema';
-import { DOMUtils } from '@jupyterlab/apputils';
+import { DOMUtils, WidgetTracker } from '@jupyterlab/apputils';
 import { IStateDB } from '@jupyterlab/statedb';
 import {
   Button,
+  ContextMenuSvg,
   LabIcon,
   caretDownIcon,
   caretRightIcon,
@@ -32,6 +36,9 @@ import {
 } from '@/src/shared/icons';
 import { ILeftPanelClickHandlerParams } from '@/src/workspace/panels/leftpanel';
 import { LegendItem } from './legendItem';
+import { ContextMenu, Menu } from '@lumino/widgets';
+import { ITranslator, nullTranslator } from '@jupyterlab/translation';
+import { rasterSubMenu, vectorSubMenu } from '../../menus';
 
 const LAYER_GROUP_CLASS = 'jp-gis-layerGroup';
 const LAYER_GROUP_HEADER_CLASS = 'jp-gis-layerGroupHeader';
@@ -50,8 +57,250 @@ interface IBodyProps {
   layerTree: IJGISLayerTree;
 }
 
+function createContextMenu(commands: CommandRegistry, model: IJupyterGISModel, translator?: ITranslator) {
+  translator = translator ?? nullTranslator;
+
+  const GIS_LAYER_ITEM =
+    '.jp-gis-layerItem:not(.jp-gis-layerGroup)';
+
+  const gisContextMenu = new ContextMenuSvg({ commands });
+
+  // LAYERS and LAYER GROUPS context menu
+  gisContextMenu.addItem({
+    command: CommandIDs.symbology,
+    selector: GIS_LAYER_ITEM,
+    rank: 1,
+  });
+
+  // Separator
+  gisContextMenu.addItem({
+    type: 'separator',
+    selector: GIS_LAYER_ITEM,
+    rank: 1.5,
+  });
+
+  gisContextMenu.addItem({
+    command: CommandIDs.removeSelected,
+    selector: GIS_LAYER_ITEM,
+    rank: 3,
+  });
+  gisContextMenu.addItem({
+    command: CommandIDs.removeSelected,
+    selector: '.jp-gis-layerGroupHeader',
+    rank: 3,
+  });
+
+  gisContextMenu.addItem({
+    command: CommandIDs.renameSelected,
+    selector: GIS_LAYER_ITEM,
+    rank: 4,
+  });
+  gisContextMenu.addItem({
+    command: CommandIDs.renameSelected,
+    selector: '.jp-gis-layerGroupHeader',
+    rank: 4,
+  });
+
+  gisContextMenu.addItem({
+    command: CommandIDs.duplicateSelected,
+    selector: GIS_LAYER_ITEM,
+    rank: 5,
+  });
+
+  gisContextMenu.addItem({
+    command: CommandIDs.zoomToLayer,
+    selector: GIS_LAYER_ITEM,
+    rank: 6,
+  });
+
+  const moveSelectedSubmenu = new Menu({ commands });
+  moveSelectedSubmenu.title.label = translator
+    .load('jupyterlab')
+    .__('Move Selection to Group');
+  moveSelectedSubmenu.id = 'jp-gis-contextmenu-movelayer';
+
+  gisContextMenu.addItem({
+    type: 'submenu',
+    selector: GIS_LAYER_ITEM,
+    rank: 7,
+    submenu: moveSelectedSubmenu,
+  });
+
+  gisContextMenu.opened.connect(() =>
+    buildGroupsMenu(gisContextMenu, model),
+  );
+
+  gisContextMenu.addItem({
+    command: CommandIDs.zoomToLayer,
+    selector: GIS_LAYER_ITEM,
+    rank: 7,
+  });
+
+  gisContextMenu.addItem({
+    command: CommandIDs.toggleDrawFeatures,
+    selector: GIS_LAYER_ITEM,
+    rank: 8,
+  });
+
+  // Separator
+  gisContextMenu.addItem({
+    type: 'separator',
+    selector: GIS_LAYER_ITEM,
+    rank: 8.5,
+  });
+
+  // Create the Download submenu
+  const downloadSubmenu = new Menu({ commands: commands });
+  downloadSubmenu.title.label = translator.load('jupyterlab').__('Download');
+  downloadSubmenu.id = 'jp-gis-contextmenu-download';
+
+  downloadSubmenu.addItem({
+    command: CommandIDs.downloadGeoJSON,
+  });
+
+  // Add the Download submenu to the context menu
+  gisContextMenu.addItem({
+    type: 'submenu',
+    selector: GIS_LAYER_ITEM,
+    rank: 9,
+    submenu: downloadSubmenu,
+  });
+
+  // Create the Processing submenu
+  const processingSubmenu = new Menu({ commands });
+  processingSubmenu.title.label = translator
+    .load('jupyterlab')
+    .__('Processing');
+  processingSubmenu.id = 'jp-gis-contextmenu-processing';
+
+  // Clip sub-submenu — groups all clipping operations
+  const clipSubmenu = new Menu({ commands });
+  clipSubmenu.title.label = translator.load('jupyterlab').__('Clip By');
+  clipSubmenu.id = 'jp-gis-contextmenu-clip';
+
+  for (const processingElement of ProcessingMerge) {
+    if (processingElement.type === 'clip') {
+      clipSubmenu.addItem({
+        command: `jupytergis:${processingElement.name}`,
+      });
+    } else {
+      processingSubmenu.addItem({
+        command: `jupytergis:${processingElement.name}`,
+      });
+    }
+  }
+
+  processingSubmenu.addItem({ type: 'separator' });
+  processingSubmenu.addItem({ type: 'submenu', submenu: clipSubmenu });
+
+  gisContextMenu.addItem({
+    type: 'submenu',
+    selector: GIS_LAYER_ITEM,
+    rank: 10,
+    submenu: processingSubmenu,
+  });
+
+  const newLayerSubMenu = new Menu({ commands });
+  newLayerSubMenu.title.label = translator.load('jupyterlab').__('Add Layer');
+  newLayerSubMenu.id = 'jp-gis-contextmenu-addLayer';
+
+  newLayerSubMenu.addItem({
+    type: 'submenu',
+    submenu: rasterSubMenu(commands),
+  });
+  newLayerSubMenu.addItem({
+    type: 'submenu',
+    submenu: vectorSubMenu(commands),
+  });
+
+  // Separator
+  gisContextMenu.addItem({
+    type: 'separator',
+    selector: GIS_LAYER_ITEM,
+    rank: 10.5,
+  });
+
+  gisContextMenu.addItem({
+    type: 'submenu',
+    selector: GIS_LAYER_ITEM,
+    rank: 11,
+    submenu: newLayerSubMenu,
+  });
+
+  return gisContextMenu;
+}
+
+
+/**
+ * Populate submenu with current group names
+ */
+function buildGroupsMenu(
+  contextMenu: ContextMenu,
+  model: IJupyterGISModel,
+) {
+  const submenu =
+    contextMenu.menu.items.find(
+      item =>
+        item.type === 'submenu' &&
+        item.submenu?.id === 'jp-gis-contextmenu-movelayer',
+    )?.submenu ?? null;
+
+  // Bail early if the submenu isn't found
+  if (!submenu) {
+    return;
+  }
+
+  submenu.clearItems();
+
+  // need a list of group name
+  const layerTree = model.getLayerTree();
+  const groupNames = getLayerGroupNames(layerTree);
+
+  function getLayerGroupNames(layerTree: IJGISLayerItem[]): string[] {
+    const result: string[] = [];
+
+    for (const item of layerTree) {
+      // Skip if the item is a layer id
+      if (typeof item === 'string') {
+        continue;
+      }
+
+      // Process group items
+      if (item.layers) {
+        result.push(item.name);
+
+        // Recursively process the layers of the current item
+        const nestedResults = getLayerGroupNames(item.layers);
+        // Append the results of the recursive call to the main result array
+        result.push(...nestedResults);
+      }
+    }
+
+    return result;
+  }
+
+  submenu.addItem({
+    command: CommandIDs.moveSelectedToGroup,
+    args: { label: '' },
+  });
+
+  groupNames.forEach(name => {
+    submenu.addItem({
+      command: CommandIDs.moveSelectedToGroup,
+      args: { label: name },
+    });
+  });
+
+  submenu.addItem({
+    command: CommandIDs.moveSelectedToNewGroup,
+  });
+}
+
 export const LayersBodyComponent: React.FC<IBodyProps> = props => {
   const model = props.model;
+  const commands = props.commands;
+
+  const layerContextMenu = createContextMenu(commands, model);
 
   const [layerTree, setLayerTree] = useState<IJGISLayerTree>(
     props.layerTree || [],
@@ -59,9 +308,16 @@ export const LayersBodyComponent: React.FC<IBodyProps> = props => {
 
   const notifyCommands = () => {
     // Notify commands that need updating
-    props.commands.notifyCommandChanged(CommandIDs.identify);
-    props.commands.notifyCommandChanged(CommandIDs.temporalController);
-    props.commands.notifyCommandChanged(CommandIDs.toggleDrawFeatures);
+    commands.notifyCommandChanged(CommandIDs.identify);
+    commands.notifyCommandChanged(CommandIDs.temporalController);
+    commands.notifyCommandChanged(CommandIDs.toggleDrawFeatures);
+  };
+
+  const _onContextMenu = (e: React.MouseEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+
+    layerContextMenu.open(e.nativeEvent);
   };
 
   const _onDragOver = (e: React.DragEvent) => {
@@ -202,7 +458,7 @@ export const LayersBodyComponent: React.FC<IBodyProps> = props => {
   }, [props.layerTree]);
 
   return (
-    <div id="jp-gis-layer-tree" onDrop={_onDrop} onDragOver={_onDragOver}>
+    <div id="jp-gis-layer-tree" onDrop={_onDrop} onDragOver={_onDragOver} onContextMenu={_onContextMenu}>
       {layerTree.map(layer =>
         typeof layer === 'string' ? (
           <LayerComponent
@@ -635,7 +891,6 @@ const LayerComponent: React.FC<ILayerProps> = props => {
   return (
     <div
       className={`${LAYER_ITEM_CLASS} ${LAYER_CLASS}
-                  ${isStorySegmentLayer ? 'jp-gis-storySegmentLayer' : ''}
                   ${selected ? ' jp-mod-selected' : ''}`}
       draggable={true}
       onDragStart={Private.onDragStart}
